+++ src/include/utils/lsyscache.h
@@ -6,7 +6,7 @@
  *
  * Copyright (c) 1994, Regents of the University of California
  *
- * $Id: lsyscache.h,v 1.8 1998/01/24 22:50:48 momjian Exp $
+ * $Id: lsyscache.h,v 1.9 1998/02/10 16:04:38 momjian Exp $
  *
  *-------------------------------------------------------------------------
  */
@@ -21,7 +21,7 @@ extern char *get_attname(Oid relid, AttrNumber attnum);
 extern AttrNumber get_attnum(Oid relid, char *attname);
 extern Oid	get_atttype(Oid relid, AttrNumber attnum);
 extern bool get_attisset(Oid relid, char *attname);
-extern int get_atttypmod(Oid relid, AttrNumber attnum);
+extern int16 get_atttypmod(Oid relid, AttrNumber attnum);
 extern RegProcedure get_opcode(Oid opid);
 extern char *get_opname(Oid opid);
 extern bool op_mergesortable(Oid opid, Oid ltype, Oid rtype,
